Adds support for more operators
This also includes a change in the SQL output: all binary expression
will surround themselves with parens (e.g. (1 + 2)). This has no
effect on the sql, but prevents accidental ambiguity when nesting
operators that have different precendence.

It also puts all operators in one namespace: { ops }. This is mostly
because I don't want people to have to keep pulling different
operators out of the fragmentum namespace and into their global
namespace.

That said, I'm considering putting all fragments / expressions into
namespaces for this exact reason. The problem there is I really like
how fragment-based queries look right now (as opposed to
builder-based), so I'll leave that decision until after I dogfood the
library for a bit.

Speaking of which, it's almost ready for dogfooding! I'm going to add
group-by and limit/offset, and then it'll have basically everything
except for non-inner joins and HAVING, neither of which I use enough
to consider it a priority. I'll probably find some issues with the
binds system, but those can be fixed pretty easily.

0.2 here we come! :tada: :smile: